[FIX] base: prevent contact name and info from being cut
This commit aims to prevent the name and contact information of a contact
from being cut short within the form view.

Prior to this commit, the width of the name and information were defined
using `w-*` utility classes, which was working well in desktop but was
causing a misalignment with the other inputs of the form view in mobile,
as well as cutting the labels short.

To prevent this issue, we simply replicate what has been done in other
module and apply these utility classes only above `medium` devices.

task-4793732