[FIX] hr_holidays: Error when selecting time off from a different timezone
When selecting an absence for a future date that falls in a different timezone (this can happen in the case of Spain), the default processing of the start and end dates may be using an incorrect timezone.

The correct approach would be to base it on the timezone of the start/end dates selected in the calendar.