[FIX] website_blog: prevent last blog post cover edition
Scenario:
- insert the last blog posts widget in a page
- open editor and edit (eg. Filter Intensity) the cover of a post
- save

Result: the change is lost

Cause: the content of the widget is dynamic, and we delete the content
in cleanForSave (that call the destroy of the widget) before saving, so
the change are not saved.

Fix: make the dynamically added cover widget unselectable, the cover of
the blog post can still be changed in other locations (blog post list,
blog post page).

opw-4633287