[IMP] web: group quick create in list
This commit adds the functionnality to quick create groups in lists
grouped by a many2one field. The button to do so is added to the footer
when possible and behaves similarly to kanban's column quick create.
It is possible to disable this feature with the group_create arch
attribute (as is done in kanban).

Part of task-4613135