[IMP] sale_project: allow project creation without service products
Previously, a project could only be created if the sale order
had a service-type product.

Now, a project can also be created from a sale order with only type
products. In this case, the project is linked to the sale order,
but not linked to any sale order line.

task-4755568