[IMP] {*_}loyalty: add new reward type: fixed price
Allows to set a fixed unit price for products, a discount line will be
created with the excess amount based on the product prices.

This allows promotions such as "Buy 2 pizzas for €15, or 5 for €30", in
this case, a first reward, requiring 2 points, with a unit price of
€7.50, and a second reward, requiring 5 points, with a unit price of
€6.00.

Due to the fact that rewards are agnostics of where the points come
from, and to avoid more bug-prone code due to a more complicate code to
be able to cover choosing which product is or isn't covered by that
reward, it was decided that these rewards would either apply on all the
products covered by that unit price, or none at all. So, in that previous
case, a third pizza would also be €7.50, and if your order is enough to
have the next step of the promotion, then all your pizzas would be at
€6.00, even if not a multiple of 5.

It was also decided, for similar reasons and to have a consistent,
predictable behavior, to only allow other discounts if they are applied on
the order and not support another discount on the same specific lines. This
is due to the fact that this fixed discount works in an reverse way,
compared to the actual ones. Where the previous discounts are always either
the same amount (in per point discount) or the same percentage, and thus
not depending on the unit price of a product, this fixed discount vary for
each product having a different unit price.

For these reasons, we are now tracking which sale order lines were
discounted and on which discounts, which will also (coincidentally) have
the benefit to ease the tax recomputation at invoicing time.

task-4433191