[IMP] stock_picking_batch: add action for merging batch/wave transfers
This commit enables merging batch/wave transfers of the same operation
type. This can be done by selecting multiple transfers in the list view
and clicking on "Merge" from the actions menu. This action merges all
the operations of the selected transfers in one of them and deletes the
others.

Task-4443804