[IMP] web, *: change xxl breakpoint for the chatter
The goal is to keep the chatter aside the formview on more devices
(small laptop, typically 13" and under).

The breakpoints are now bootstrap default since we removed the custom
value on the `xxl` breakpoint.

`VSM` value was not relevant anymore, bootstrap at the time didn't have
as much breakpoints, which probably required this custom breakpoint for
mobile devices. Now the `SM` value is the default value used for these
viewports.

task-4806037