[IMP] mail,*: chatter live update
The point of this commit is to have a live chatter. Therefore you will
receive new messages/followers update/attachments/activities directly in
your chatter without refreshing the page or doing an action yourself,
allowing you to have a more lively discussion with people connected to
the same chatter.

This is mainly done by inheriting `bus.listener.mixin` in `mail.thread`
which allows to send bus notification to all clients listening to one of
the subchannels of the model.

The discuss app (which was already supporting those bus features) is
relieved of some actions that are now
delegated to the mail.thread directly.

In more detail, each model inheriting from `mail.thread` has a
corresponding channel to subscribe to (i.e.
`thread-{model_name}_{model_id}`). On top of that an `internal`
subchannel is automatically assigned to the user depending on whether
the user is internal or not. This allows safe sharing of `Log Notes` as
well as some other internal notifications not shared with the external
world.

The user is subscribed to the channel whenever the chatter is mounted in
the DOM or when the thread is changed.

task-3660221