[FIX] website: ignore mutations on hide/show a modal popup in builder
Some of the mutation happening when toggling the visibility of a popup
were tracked in the history. Thus, when toggling the visibility of a
popup and doing undo/redo, the actual visibility of a popup could be
inconsistent with the state of the toggle button.

To avoid adding mutations in the history when toggling visibility,
`ignoreDOMMutations` is used because a few changes are made by bootstrap
in attributes that are commonly used. Bootstrap may also delay the
hiding of a popup, so it also needs to be used in the method that does
the hiding

task-4367641