[FIX] html_builder, website: avoid previewing the same action twice
1) Avoid double previews

On `BuilderSelectItem`s, the action is previewed on mouseover and on
focusin. This means that when using the mouse and clicking on an item,
there are always 2 previews before committing (on click). This commit
blocks the second preview if the current item is already being
previewed.

In addition, we also make sure that the test tests the actual dropdown
item, and not the component marked with `ignoreBuilderItem`. By
targeting the proper item, we ensure that all events (hover, focusin,
click) are dispatched: the ignored one being in a `.d-none` container,
the `focusin` event was not dispatched.

2) Await revert before next preview

The commit also makes sure that, after an async preview of action A, the
next preview of action B first awaits action A's revert. Otherwise,
depending on how long each preview takes, we could end up with the wrong
action applied.

This PR follows [the refactor into html_builder].

[the refactor into html_builder]: https://github.com/odoo/odoo/commit/9fe45e2b7ddb

Related to task-4367641