[IMP] account, accountant: rework of To Check
The To Check feature should only be used by the accountants. A random user who only use accounting to create bills and invoices shouldn't be allowed to review and mark a move as checked.

This is why now we only show the feature to accountant (Bookkeeper and Accounting admins) and raise an error if a user without the good rights try to change the checked field.

task-4807991