[IMP] mail, calendar, crm: add activity reschedule mixin widget
A new widget is added to the CRM lead list view, replacing the previous
"Snooze 7d" option. This widget will function similarly to the "reschedule"
widget already present on mail.activity views, and works for all models
inheriting from mail.activity.mixin.

If the Calendar module is installed, and the next activity linked to
the record is a Meeting, the reschedule button is replaced by a action
button to reschedule the meeting, similarly to pre-existing
behavior on the CRM view.

task-4800833