[FIX] website_slides: fix broken share link in email
Steps to reproduce
===================
1. Add new content to the course and publish it.
2. A notification email is sent out to attendees of the course.
3. Click the 'View Content' button in the email.
It leads to a 404 error page.

With https://github.com/odoo/odoo/commit/074be1ed36891065361a3cf40f8e189ba5fccb15, we replaced the ID-based route with website_absolute_url,
which uses a slugified URL format, and we don't have a slugify route with `/share`.

This commit adds a new route to make it compatible with `/share`.

Task-4797324