[IMP] registry: improve performance of model setup during install
While loading the registry, we call `_setup` on every model.  This
operation is slow when many models are present in the registry.

This is more visible during install/updates of modules, where the setup
is called for each module and each model, even when the model was not
impacted by the current module.

This commit avoid to re-setup models that are not modified, trying to
keep the parts of the registry that should still be valid.

This approach is obviously more complex than the previous one, since we
need to find all possible cases that could lead to a model being
modified, but the performance improvement is worth it.

This commit could lead to a 15~30% improvement.
 (To confirm with the final version)

This number is extracted from this build
https://runbot.odoo.com/runbot/batch/1960315/build/81155809
where the install time was 12 minutes, when it usually takes around
14-25 minutes.

To be more precisen the time spent in setup_model_classes was around
5-5:30 minutes before this change, and arround 1:20 minutes after.