[IMP] web: do not distribute negation on operators
A domain of the form ["!", (p, ">", v)]
is not always equivalent to the domain
[(p, "<=", v)].
The equivalence only hold in specific case and
depends on the type/complexity of the path p.
For instance if p is a datetime field,
the first domain captures records where p is less
or equal to v or p is not set while the second
domain captures only records where p is less or
equal to v.
More generally, distribute "!" on the operators and
"invert" them is not always valid. So for simplicity,
we stop doing it completely.