[FIX] sale: mark orders as sent on scheduled mails
Versions
--------
- 18.0

Fixed in 18.1+ via 752fdb2d2718

Steps
-----
1. Create a quotation;
2. click "Send by Email";
3. schedule the message to get sent at a later time;
4. wait for the message to send, or click "Send Now" in the chatter;
5. refresh the page.

Issue
-----
Quotation state did not get set to `sent`.

Cause
-----
Before this commit, marking the quotation as `sent` happened via a
context value. This worked for earlier versions, but with the
introduction of scheduled messaging, the context is no longer available
when the scheduled message gets sent.

Solution
--------
Add the `mark_so_as_sent` context value to the `notification_parameters`
field of `mail.scheduled.message`, and check for this value in the
`_post_message` method that gets called when sending the message. If
present, re-introduce the context value in the call to `super`.

opw-4794010