[FIX] web: fix highlight text in settings
This commit removes the use of odoomark in highlightText within the
settings module, as it was unintentionally inserting excessive <br> tags.
Some spacing issues are also handled in consequence to this change.
It also restores bold styling for search result highlights which was
using <b> tags before but was removed with the new highlight system.