[IMP] website: add description input field in map snippet
Issue:
Users are unable to edit the description immediately after adding a map
snippet.
They must first click elsewhere and then return to the description field
to make changes.

Fix:
Add an input field in options of map snippet.
This input field is visible when description checkbox in true.
Any Update in the field will update the description of the snippet.

task-4604018