[IMP] crm: add rotting fields
New fields are added to the stage and lead models, allowing to get a
better sense of when an opportunity gets stale. When an opportunity
spends too long without seeing activity (record updates, chatter
messages, completed activities)..., it will appear in a different
color in kanban views, to highlight it and bring it to attention.

task-4684660