[IMP] mrp: introduce Standard Batch Size on BOM test case
Added a test case to validate the behaviour of the Standard Batch Size feature
on Bill of Materials. The test ensures that product quantities are correctly
split according to the batch size defined on the BOM, and that edge cases like
partial batches are handled properly.

Task ID: [4395548](https://www.odoo.com/odoo/project/966/tasks/4395548)