[IMP] html_builder: allow to activate an element after using an option
This commit allows to add information in the current step, that will be
used when the step will be added, in order to activate the containers of
a specific element. This information will need to be added during the
option execution, by using the `setStepExtra` shared function of the
`history` plugin.

The information that can be added is:
- `nextTarget`: the element of which we want to activate the containers.
- `forceContainerUpdate`: a boolean used when we want to update the
containers of the current target. Indeed, there is some code preventing
to update them if they did not change, but in some cases, we still want
to do it.

See next commits of this PR for real examples.

task-4367641