[FIX] web: SelectCreateDialog: allow click on rows
In a SelectCreateDialog, select a record by clicking on the
checkbox. From this point, we entered the selection mode. Before
this commit, clicking on a row (not its checkbox), even in that
mode would select that record only and close the dialog, ignoring
previously selected records. From now on, once in selection mode,
more records can be selected by clicking on their row, not only
their checkbox. This prevents a misclick from closing the dialog,
when trying to select several records.

task~4830114