[FIX] website: hide parallax option in dialog
Steps to Reproduce :

- Enter edit mode.
- Drag and drop the s_popup snippet.
- The parallax option is visible in the background options.

Previously, the parallax option was visible when editing the background
image of a snippet inside a dialog, even though scrolling is not
possible in that context. This didn’t make sense, so this commit hides
the parallax option when the editing element is within a dialog.

This commit also move the parallax option above the shape background
options instead of below if both bg image and bg shape are applied.

The option also has been renamed from 'Parallax' to 'Scroll Effect' as
it was done before the mysterious-egg project.