[FIX] purchase: remove redundant uom field from purchase view
This commit removes the unnecessary copy of `product_uom_id` field on
`purchase_order_line` list view in `purchase_order` form view.
Previously, the view had the same field twice: one invisible in case the
user doesn't have the uom group, and another one only visible if the
user has the group. The first one is not needed, because the view
automatically loads the field if it's needed as a dependency in any
other field. This also led to an issue with "Studio", where a user was
not able to change the column label, because the name change only
affected the first occurrence of the field (which was the invisible
one).

Task-4818427