[FIX] account: fix _search_placeholder_code
account.account.placeholder_code is seen as groupable by the webclient since there is an implementation inside _field_to_sql.
But groupby it, leads to an UserError because the search method doesn't implement equality (needed since extra_domain returns by formatted_read_group contains leaf like ('placeholder_code', '=', ...)).

Add the implementation for the equality in the search method. The test will be added in the Fix it. A test that target every model will be added in master on the PR that discovered the issue https://github.com/odoo/odoo/pull/133224.