[FIX] website: review carousel options and clean the code
This commit fixes various things about the carousel options:

- Remove the normalize resource: the issue it was fixing does in fact
not happen, making this code useless.

- Hide the carousel controllers when removing a slide, if there is only
one slide left.

- Put all the code concerning the addition of a slide in the `addSlide`
function instead of using the clone resources, making it easier to
follow the flow.

- Not use the `remove` plugin `removeElement` shared function when
removing a slide and its indicator, as we do not want to go up to the
parent and so on, we simply want to delete the element.

- Remove the hack of "adding a step and then updating the containers" to
activate the active slide, as it is not correct and will add multiple
steps for one operation. Instead, add the needed information in the
current step, as explained in the previous commit of this PR. It allows
to always have the right slide counter, no matter how much we undo/redo.

- Add some comments and docstrings.

- Use better variable names (to uniformize and simplify them).

- Clean the code.

task-4367641