[IMP] crm: clarify team member form view
Previously, on the CRM team member form view, the Capacity section was
presented without a label and without particular highlight. It was
unclear that the maximum capacity could be edited on this view, and
made for a jarring visual.

This commit updates the view by adding a label, and adding a permanent
highlight on the input field.

Additionally, it updates the help sentences for lead day counts and
lead month counts so that they're proper sentences.

(to squash into the planned further team member changes?)

task-4684660