[IMP] orm: iterate over model classes
The current implementation of setup iterates over env (models) instead
of registry (models classes), even if in most cases the class will be
needed. This is

This is not efficient and leads to many call to __get_item__.

This commit changes the iteration to use the registry directly, and use
the model classes, instantiating it with an env when needed. Note that
in most case this env could be removed by replacing some methods by
class methods or hard-coded logic.

This work was initially made in #211658 but was extracted here.