[IMP] website_partner: add replace image option on partner page
Before this commit:

After creating a partner in the Contacts module, if the user publishes
the partner's page and then realizes they forgot to add an image,
there's no way to upload or update the image directly from the page
editing view. This creates unnecessary friction, as the user must return
to the backend to make the change.

After this commit:

Users can add an image from the website page, just like for products.

task-4760181