[FIX] point_of_sale: load account_move model
Steps to reproduce
- install `l10n_pe_edi_pos` module
- open a pos session
- create a order, when checking out, tick `invoice` checkbox
- pay and complete the order
- go to order menu, from top-right hamburger  button
- select the previously completed order, apply for refund
Observation: you are navigated to Order screen
Expectation: A dialog for reason should be open first

Issue
- after this commit odoo/odoo@96ee0b6 and odoo/enterprise@e3e84bb account_move
field is used instead of `invoiced` state of order. But the `account_move` field
is always undefined.
This is due to fact that when pos_order is fulfilled from `syncAllOrders`
function, a call to `missingRecursive` is made, as `account.move` model is never
 loaded, it is unable to link it to appropriate model and results to undefined

Fix
- load `account.move` model at init

opw-4744785