[FIX] website,*: move is-ready to iframe body
The goal of this commit is to check whether a website page is ready
or not in the backend (iframe) or frontend. We'll therefore move
the is-ready attribute to the iframe body.
This attribute will be added when interactions are ready.

In this commit we're going to set a few intéderministes tours that need
to wait for interactions to be ready before they can work.