[IMP] base: move profiler code in qweb
The decorator is adds an overhead to each call and makes it harder to
debug. These are already called during initialization of the module so
qweb is aware of profiling and profiling arguments must match the only
decorated method. Therefore, we can inline the profiling code where it
is needed.